Move google coordinator to separate module
